Vytvorenie postupného dedenia
