+++ src/Model/EdgeDAO.cpp
@@ -140,7 +140,7 @@ bool Model::EdgeDAO::addEdgesColorToDB( QSqlDatabase* conn, QMap<qlonglong, osg:
 
 	while ( iEdges != edges->constEnd() ) {
 		//ulozime farbu len hranam, ktore maju farbu inu nez default
-		if ( iEdges.value()->getEdgeColor().r() != 1.f || iEdges.value()->getEdgeColor().g() != 1.f ||iEdges.value()->getEdgeColor().b() != 1.f ||iEdges.value()->getEdgeColor().a() != 1.f ) {
+        if ( iEdges.value()->getEdgeColor()[0] != 1.f || iEdges.value()->getEdgeColor()[1] != 1.f ||iEdges.value()->getEdgeColor()[2] != 1.f ||iEdges.value()->getEdgeColor()[3] != 1.f ) {
 			if ( meta ) {
 				if ( newMetaEdgeID.contains( iEdges.value()->getId() ) ) {
 					edgeIdIter = newMetaEdgeID.find( iEdges.value()->getId() );
@@ -154,10 +154,10 @@ bool Model::EdgeDAO::addEdgesColorToDB( QSqlDatabase* conn, QMap<qlonglong, osg:
 				edgeID = iEdges.value()->getId();
 			}
 
-			addSetings( conn, iEdges.value()->getGraph()->getId(), layout->getId(), edgeID, "color_r", iEdges.value()->getEdgeColor().r() );
-			addSetings( conn, iEdges.value()->getGraph()->getId(), layout->getId(), edgeID, "color_g", iEdges.value()->getEdgeColor().g() );
-			addSetings( conn, iEdges.value()->getGraph()->getId(), layout->getId(), edgeID, "color_b", iEdges.value()->getEdgeColor().b() );
-			addSetings( conn, iEdges.value()->getGraph()->getId(), layout->getId(), edgeID, "color_a", iEdges.value()->getEdgeColor().a() );
+            addSetings( conn, iEdges.value()->getGraph()->getId(), layout->getId(), edgeID, "color_r", iEdges.value()->getEdgeColor()[0]  );
+            addSetings( conn, iEdges.value()->getGraph()->getId(), layout->getId(), edgeID, "color_g", iEdges.value()->getEdgeColor()[1]  );
+            addSetings( conn, iEdges.value()->getGraph()->getId(), layout->getId(), edgeID, "color_b", iEdges.value()->getEdgeColor()[2]  );
+            addSetings( conn, iEdges.value()->getGraph()->getId(), layout->getId(), edgeID, "color_a", iEdges.value()->getEdgeColor()[3]  );
 		}
 
 		++iEdges;
