+++ src/Model/GraphDAO.cpp
@@ -215,11 +215,16 @@ Data::Graph* Model::GraphDAO::getGraph( QSqlDatabase* conn, bool* error2, qlongl
 			newGraph->addEdge( edgeID, edgeName, iNodes1.value(), iNodes2.value(), type, isOriented );
 
 			if ( edgeColors.contains( edgeID ) ) {
+                cv::Vec4f color;
+                color[0]=edgeColors.value( edgeID ).r();
+                color[1]=edgeColors.value( edgeID ).g();
+                color[2]=edgeColors.value( edgeID ).b();
+                color[3]=edgeColors.value( edgeID ).a();
 				if ( newGraph->getEdges()->contains( edgeID ) ) {
-					newGraph->getEdges()->find( edgeID ).value()->setEdgeColor( edgeColors.value( edgeID ) );
+                    newGraph->getEdges()->find( edgeID ).value()->setEdgeColor( color );
 				}
 				else {
-					newGraph->getMetaEdges()->find( edgeID ).value()->setEdgeColor( edgeColors.value( edgeID ) );
+                    newGraph->getMetaEdges()->find( edgeID ).value()->setEdgeColor( color );
 				}
 			}
